Public

Library

Index

SAASGP4.DTGToUTC

SAASGP4.astroFuncGetInfo

SAASGP4.astroFuncInit

SAASGP4.closeLogFile

SAASGP4.envGetFkConst

SAASGP4.envGetFkIdx

SAASGP4.envGetFkPtr

SAASGP4.envGetGeoConst

SAASGP4.envGetGeoIdx

SAASGP4.envGetGeoStr

SAASGP4.envGetInfo

SAASGP4.envInit

SAASGP4.envLoadFile

SAASGP4.envSaveFile

SAASGP4.envSetFkIdx

SAASGP4.envSetGeoIdx

SAASGP4.envSetGeoStr

SAASGP4.getInitDllNames

SAASGP4.getLastErrMsg

SAASGP4.getLastInfoMsg

SAASGP4.mainGetInfo

SAASGP4.mainInit

SAASGP4.openLogFile

SAASGP4.sgp4GetInfo

SAASGP4.sgp4Init

SAASGP4.sgp4InitSat

SAASGP4.sgp4PropDs50UTC

SAASGP4.sgp4PropDs50UTC!

SAASGP4.sgp4PropMse

SAASGP4.sgp4PropMse!

SAASGP4.sgp4RemoveAllSats

SAASGP4.sgp4RemoveSat

SAASGP4.sgp4SetLicFilePath

SAASGP4.timeFuncGetInfo

SAASGP4.timeFuncInit

SAASGP4.tleAddSatFrLines

SAASGP4.tleGetInfo

SAASGP4.tleInit

SAASGP4.tleRemoveAllSats

SAASGP4.tleRemoveSat

Interface

SAASGP4.DTGToUTC — Method.

DTGToUTC(dtg)

Converts a time in one of the DTG formats to a time in ds50UTC. DTG15, DTG17, DTG19, and DTG20 formats are accepted

source

SAASGP4.astroFuncGetInfo — Method.

astroFuncGetInfo()

Returns information about the AstroFunc DLL

source

SAASGP4.astroFuncInit — Method.

astroFuncInit(mainHandle)

Initializes AstroFunc DLL for use in the program

source

SAASGP4.closeLogFile — Method.

closeLogFile()

Closes the currently open log file

source

SAASGP4.envGetFkConst — Method.

envGetFkConst(id)

Retrieves the value of one of the constants from the current fundamental catalogue (FK) model

source

SAASGP4.envGetFkIdx — Method.

Returns the current fundamental catalogue (FK) setting

source

SAASGP4.envGetFkPtr — Method.

envGetFkPtr()

Returns a handle that can be used to access the fundamental catalogue (FK) data structure

source

SAASGP4.envGetGeoConst — Method.

envGetGeoConst(id)

Retrieves the value of one of the constants from the current Earth constants (GEO) model

source

SAASGP4.envGetGeoIdx — Method.

Returns the current Earth constants (GEO) setting

source

SAASGP4.envGetGeoStr — Method.

envGetGeoStr()

Returns the name of the current Earth constants (GEO) model

source

SAASGP4.envGetInfo — Method.

envGetInfo()

Returns information about the EnvConst DLL

source

SAASGP4.envInit — Method.

envInit(mainHandle::Clonglong)

Initializes the EnvInit DLL for use in the program

source

SAASGP4.envLoadFile — Method.

envLoadFile(path)

Reads Earth constants (GEO) model and fundamental catalogue (FK) model settings from a file

source

SAASGP4.envSaveFile — Function.

envSaveFile(path, append = 0, format = 0)

Saves the current Earth constants (GEO) model and fundamental catalogue (FK) model settings to a file

source

SAASGP4.envSetFkIdx — Method.

envSetFkIdx(id)

Changes the fundamental catalogue (FK) setting to the specified value

source

SAASGP4.envSetGeoIdx — Method.

envSetGeoIdx(id)

Changes the Earth constants (GEO) setting to the specified value

source

SAASGP4.envSetGeoStr — Method.

envSetGeoStr(geoStr)

Changes the Earth constants (GEO) setting to the model specified by a string literal

source

SAASGP4.getInitDllNames — Method.

getInitDllNames()

Returns a list of names of the Standardized Astrodynamic Algorithms DLLs that were initialized successfully

source

SAASGP4.getLastErrMsg — Method.

getLastErrMsg()

Returns a character string describing the last error that occurred

source

SAASGP4.getLastInfoMsg — Method.

getLastInfoMsg()

Returns a character string describing the last informational message that was recorded

source

SAASGP4.mainGetInfo — Method.

mainGetInfo()

Returns information about the DllMain DLL

source

SAASGP4.mainInit — Method.

mainInit()

Returns a handle which can be used to access the static global data set needed by the Standardized Astrodynamic Algorithms DLLs to communicate among themselves

source

SAASGP4.openLogFile — Method.

openLogFile(path)

Opens a log file and enables the writing of diagnostic information into it

source

SAASGP4.sgp4GetInfo — Method.

sgp4GetInfo()

Returns information about the Sgp4 DLL

source

SAASGP4.sgp4Init — Method.

sgp4Init(mainHandle)

Initializes the Sgp4 DLL for use in the program

source

SAASGP4.sgp4InitSat — Method.

sgp4InitSat(satkey)

Initializes an SGP4 satellite from an SGP or SGP4 TLE

source

SAASGP4.sgp4PropDs50UTC! — Method.

sgp4PropDs50UTC!(pos, vel, llh, mse, satkey, ds50UTC)

Propagates a satellite, represented by the satKey, to the time expressed in days since 1950, UTC

source

SAASGP4.sgp4PropDs50UTC — Method.

sgp4PropDs50UTC(satkey, ds50UTC)

Propagates a satellite, represented by the satKey, to the time expressed in days since 1950, UTC

source

SAASGP4.sgp4PropMse! — Method.

sgp4PropMse!(pos, vel, llh, ds50UTC, satkey, mse)

Propagates a satellite, represented by the satKey, to the time expressed in minutes since the satellite's epoch time. Stores the results in the provided variables.

Examples

julia> pos = Array{Cdouble, 1}(undef, 3)
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

julia> vel = Array{Cdouble, 1}(undef, 3)
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

julia> llh = Array{Cdouble, 1}(undef, 3)
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

julia> ds50UTC = Ref{Cdouble}(0.0)
Base.RefValue{Float64}(0.0)

julia> SAASGP4.sgp4PropMse!(pos, vel, llh, ds50UTC, satkey, mse)

julia> pos
3-element Array{Float64,1}:
 42155.0257401623
   980.5342672211367
    -9.307135121360039

julia> vel
3-element Array{Float64,1}:
 -0.07133558166329691
  3.07373221015029
 -0.0018402927526264539

julia> llh
3-element Array{Float64,1}:
    -0.012659361941212956
    34.46169610621692
 35788.29393717118

julia> ds50UTC[]
18314.489312984446

source

SAASGP4.sgp4PropMse — Method.

sgp4PropMse(satkey, mse)

Propagates a satellite, represented by the satKey, to the time expressed in minutes since the satellite's epoch time

source

SAASGP4.sgp4RemoveAllSats — Method.

sgp4RemoveAllSats()

Removes all currently loaded satellites from memory

source

SAASGP4.sgp4RemoveSat — Method.

sgp4RemoveSat(satkey)

Removes a satellite, represented by the satKey, from the set of satellites

source

SAASGP4.sgp4SetLicFilePath — Method.

sgp4SetLicFilePath(licFilePath)

Initializes the Sgp4 DLL for use in the program

source

SAASGP4.timeFuncGetInfo — Method.

timeFuncGetInfo()

Returns information about the TimeFunc DLL

source

SAASGP4.timeFuncInit — Method.

timeFuncInit(mainHandle)

Initializes the TimeFunc DLL for use in the program

source

SAASGP4.tleAddSatFrLines — Method.

tleAddSatFrLines(line1, line2)

Adds a TLE (satellite), using its directly specified first and second lines

source

SAASGP4.tleGetInfo — Method.

tleGetInfo()

Returns information about the Tle DLL

source

SAASGP4.tleInit — Method.

tleInit(mainHandle)

Initializes Tle DLL for use in the program

source

SAASGP4.tleRemoveAllSats — Method.

tleRemoveAllSats()

Removes all the TLEs from memory

source

SAASGP4.tleRemoveSat — Method.

tleRemoveSat(satkey)

Removes a TLE represented by the satKey from memory

source